Made some EDA in R
I made comments in the task Look at the data in Notion. Look here for more explanation. I tried to provide the documentation there. A lot of the time I looked at the dataframes and in some cases like the anti saccade i tried to plot the fixations and fixpoints